Caching objects table
Testing out a new feature for caching the objects table. I was using this proc the other day to track down the process chain of an extremely large and complicated process. I needed to run this probably 50+ times...so I thought maybe it would run a little better if those results were cached, especially on systems with lots of databases.