Change the structure to read dependent info from temp table
The idea here is this may eventually become a table valued function

This could be converted to be paired with a view that generates the values
needed for the TVF. The TVF simply structures the statements from the dependencies.

However, the scripts can also be populated from any source, not just the
system talbes.